requirements.txt
tinkoff-investments==0.2.0b117
protobuf>=4.25.1,<5
dbt-core==1.8.8
dbt-postgres==1.9.1
aiohttp==3.10.11
nest-asyncio==1.6.0
grpcio
apache-airflow==2.10.4
airflow-clickhouse-plugin==1.4.0
airflow-providers-clickhouse==0.0.1
beautifulsoup4==4.12.3
openpyxl==3.1.2
xmltodict==0.13.0
clickhouse-driver==0.2.8
SQLAlchemy==1.4.54
# psycopg2-binary
dataclasses_json==0.6.7
pandas==2.1.4
pendulum==3.0.0
kafka-python==2.2.15
python-dotenv==1.0.1
requests==2.32.0
selenium==4.27.1
